backend/requirements.txt
flask
flask-cors
bcrypt
python-dotenv
openai>=1.0.0
PyJWT
